+++ driver/web/conversions_auth.go
@@ -15,13 +15,17 @@
 package web
 
 import (
+	"context"
 	"core-building-block/core/model"
 	Def "core-building-block/driver/web/docs/gen"
 	"core-building-block/utils"
+	"encoding/base64"
 
-	"github.com/rokwire/core-auth-library-go/v2/authorization"
-	"github.com/rokwire/core-auth-library-go/v2/authservice"
-	"github.com/rokwire/core-auth-library-go/v2/authutils"
+	"github.com/lestrrat-go/jwx/jwk"
+	"github.com/rokwire/core-auth-library-go/v3/authorization"
+	"github.com/rokwire/core-auth-library-go/v3/authservice"
+	"github.com/rokwire/core-auth-library-go/v3/authutils"
+	"github.com/rokwire/core-auth-library-go/v3/keys"
 	"github.com/rokwire/logging-library-go/v2/errors"
 	"github.com/rokwire/logging-library-go/v2/logutils"
 )
@@ -64,14 +68,14 @@ func loginSessionsToDef(items []model.LoginSession) []Def.LoginSession {
 	return result
 }
 
-func pubKeyFromDef(item *Def.PubKey) *authservice.PubKey {
+func pubKeyFromDef(item *Def.PubKey) *keys.PubKey {
 	if item == nil {
 		return nil
 	}
-	return &authservice.PubKey{KeyPem: item.KeyPem, Alg: item.Alg}
+	return &keys.PubKey{KeyPem: item.KeyPem, Alg: item.Alg}
 }
 
-func pubKeyToDef(item *authservice.PubKey) *Def.PubKey {
+func pubKeyToDef(item *keys.PubKey) *Def.PubKey {
 	if item == nil {
 		return nil
 	}
@@ -339,25 +343,60 @@ func serviceScopeListToDef(items []model.ServiceScope) []Def.ServiceScope {
 	return out
 }
 
-func jsonWebKeyToDef(item *model.JSONWebKey) *Def.JWK {
+func jsonWebKeyToDef(item jwk.Key) *Def.JWK {
 	if item == nil {
 		return nil
 	}
-	return &Def.JWK{Alg: Def.JWKAlg(item.Alg), Kid: item.Kid, Kty: Def.JWKKty(item.Kty), Use: Def.JWKUse(item.Use), N: item.N, E: item.E}
-}
 
-func jsonWebKeySetDef(items *model.JSONWebKeySet) *Def.JWKS {
-	if items == nil || items.Keys == nil {
+	key := &Def.JWK{Alg: Def.JWKAlg(item.Algorithm()), Kid: item.KeyID(), Kty: Def.JWKKty(item.KeyType()), Use: Def.JWKUse(item.KeyUsage())}
+
+	switch t := item.(type) {
+	case jwk.RSAPublicKey:
+		nStr := base64.URLEncoding.EncodeToString(t.N())
+		key.N = &nStr
+
+		eStr := base64.URLEncoding.EncodeToString(t.E())
+		key.E = &eStr
+	case jwk.ECDSAPublicKey:
+		crv := t.Crv().String()
+		key.Crv = &crv
+
+		xStr := base64.URLEncoding.EncodeToString(t.X())
+		key.X = &xStr
+
+		yStr := base64.URLEncoding.EncodeToString(t.Y())
+		key.Y = &yStr
+	case jwk.OKPPublicKey:
+		crv := t.Crv().String()
+		key.Crv = &crv
+
+		xStr := base64.URLEncoding.EncodeToString(t.X())
+		key.X = &xStr
+	default:
 		return nil
 	}
-	out := make([]Def.JWK, len(items.Keys))
-	for i, item := range items.Keys {
-		defItem := jsonWebKeyToDef(&item)
-		if defItem != nil {
-			out[i] = *defItem
-		} else {
-			out[i] = Def.JWK{}
+
+	return key
+}
+
+func jsonWebKeySetDef(set jwk.Set) *Def.JWKS {
+	if set == nil {
+		return nil
+	}
+	out := make([]Def.JWK, set.Len())
+
+	ctx := context.Background()
+	for setIter := set.Iterate(ctx); setIter.Next(ctx); {
+		item := setIter.Pair()
+		if key, ok := item.Value.(jwk.Key); ok {
+			defItem := jsonWebKeyToDef(key)
+			if defItem != nil {
+				out[item.Index] = *defItem
+				continue
+			}
 		}
+
+		out[item.Index] = Def.JWK{}
 	}
 	return &Def.JWKS{Keys: out}
 }
@@ -373,9 +412,11 @@ func authTypeToDef(item *model.AuthType) *Def.AuthType {
 	if idVal != "" {
 		id = &idVal
 	}
+
+	params := item.Params
 	return &Def.AuthType{Id: id, Code: item.Code, Description: item.Description,
 		IsExternal: item.IsExternal, IsAnonymous: item.IsAnonymous, UseCredentials: item.UseCredentials,
-		IgnoreMfa: item.IgnoreMFA, Params: &Def.AuthType_Params{AdditionalProperties: item.Params}}
+		IgnoreMfa: item.IgnoreMFA, Params: &params}
 }
 
 func authTypesToDef(items []model.AuthType) []Def.AuthType {